requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
flask==1.1.1
python-dotenv == 0.13.0
gunicorn == 20.0.4
postal == 1.1.8
flask_cors == 3.0.8